Update all references to get_ram_size to unsigned long
With the recent change to use unsigned longs over longs for get_ram_size
all references have to be updated. Since the old behavior was broken to
begin with, changes are kept as close to begin with for boards. For
boards where there where no consequences and could be safely converted
from long to unsigned long, those have been changed.

+++ arch/arm/cpu/arm926ejs/mxs/spl_mem_init.c
@@ -226,7 +226,7 @@ uint32_t mxs_mem_get_size(void)
 	da = vt[4];
 	vt[4] = data_abort_memdetect_handler;
 
-	sz = get_ram_size((long *)PHYS_SDRAM_1, PHYS_SDRAM_1_SIZE);
+	sz = (uint32_t)get_ram_size((unsigned long *)PHYS_SDRAM_1, PHYS_SDRAM_1_SIZE);
 
 	/* Restore the old DABT handler. */
 	vt[4] = da;
