+++ board/netta/netta.c
@@ -380,7 +380,7 @@ phys_size_t initdram(int board_type)
 		}
 	}
 
-	size = get_ram_size((long *)0, SDRAM_MAX_SIZE);
+	size = (long int)get_ram_size((unsigned long *)0, SDRAM_MAX_SIZE);
 
 #if 0
 	printf("check 0\n");
