+++ board/samsung/smdk5250/exynos5-dt.c
@@ -115,11 +115,11 @@ int board_init(void)
 int dram_init(void)
 {
 	int i;
-	u32 addr;
+	unsigned long addr;
 
 	for (i = 0; i < CONFIG_NR_DRAM_BANKS; i++) {
 		addr = CONFIG_SYS_SDRAM_BASE + (i * SDRAM_BANK_SIZE);
-		gd->ram_size += get_ram_size((long *)addr, SDRAM_BANK_SIZE);
+		gd->ram_size += get_ram_size((unsigned long *)addr, SDRAM_BANK_SIZE);
 	}
 	return 0;
 }
@@ -244,11 +244,11 @@ int power_init_board(void)
 void dram_init_banksize(void)
 {
 	int i;
-	u32 addr, size;
+	unsigned long addr, size;
 
 	for (i = 0; i < CONFIG_NR_DRAM_BANKS; i++) {
 		addr = CONFIG_SYS_SDRAM_BASE + (i * SDRAM_BANK_SIZE);
-		size = get_ram_size((long *)addr, SDRAM_BANK_SIZE);
+		size = get_ram_size((unsigned long *)addr, SDRAM_BANK_SIZE);
 
 		gd->bd->bi_dram[i].start = addr;
 		gd->bd->bi_dram[i].size = size;
