+++ board/ve8313/ve8313.c
@@ -81,7 +81,7 @@ static long fixed_sdram(void)
 
 	/* now check the real size */
 	disable_addr_trans ();
-	msize = get_ram_size (CONFIG_SYS_DDR_BASE, msize);
+	msize = (u32)get_ram_size (CONFIG_SYS_DDR_BASE, msize);
 	enable_addr_trans ();
 #endif
